页面提交回复

+++ src/main/java/com/springboot/community/dto/QuestionDTO.java
@@ -12,13 +12,13 @@ import lombok.Data;
 @Data
 /*QuestionDTO*/
 public class QuestionDTO {
-    private Integer id;//问题id
+    private Long id;//问题id
     private String title;//问题标题
     private String description;//问题描述
     private String tag;//问题标签
     private Long gmtCreate;//创建时间
     private Long gmtModified;//修改时间
-    private Integer creator;//用户id
+    private Long creator;//用户id
     private Integer viewCount;//浏览数量
     private Integer commentCount;//阅读数量
     private Integer likeCount;//关注数量
